Commit to begin Frontend

+++ database/migrations/2024_12_06_000000_create_perfil_clinico_motociclistas_table.php
@@ -16,12 +16,12 @@ class CreatePerfilClinicoMotociclistasTable extends Migration
             $table->string('medicamentosHabituales', 255);
             $table->string('contactoEmergencia', 45);
             $table->string('telefonoEmergencia', 15);
-            $table->unsignedBigInteger('motociclistasIdMotociclistas'); // Clave foránea
-            $table->foreign('motociclistasIdMotociclistas', 'fk_motociclistas_perfil')
-                  ->references('idMotociclistas')
-                  ->on('motociclistas')
-                  ->onDelete('cascade')
-                  ->onUpdate('cascade');
+            $table->unsignedInteger('motociclistasIdMotociclistas'); // Asegúrate de que sea unsignedInteger
+        $table->foreign('motociclistasIdMotociclistas', 'fk_motociclistas_perfil')
+              ->references('idMotociclistas')
+              ->on('motociclistas')
+              ->onDelete('cascade')
+              ->onUpdate('cascade');
             $table->timestamps();
         });
